AP: Allow starting multiple interfaces within single MLD
Add support for including multiple hostapd interfaces in
the same AP MLD, i.e., all using the same underlying driver
network interface.

To do so, when a new hostapd interface is added, if there is
already another interface using the same underlying network
interface, associate the new interface with the same private
data object, instead of creating a new one.

As some of the BSS's are non first BSS's, meaning that they reuse
the drv_priv of the initial BSS, make sure not to double free it.

Currently multiple BSS entries are not supported so always use bss[0]
for MLD.

Signed-off-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>
Signed-off-by: Ilan Peer <ilan.peer@intel.com>